Move ArrayOverlapLayer to layers.py, this module has no numpy/pandas imports, for scheduler security
