+++ dask/tests/test_layers.py
@@ -5,9 +5,12 @@ import pytest
 distributed = pytest.importorskip("distributed")
 
 import sys
+from operator import getitem
 
 from distributed import Client, SchedulerPlugin
 from distributed.utils_test import cluster, loop  # noqa F401
+
+from dask.layers import fractional_slice
 
 
 class SchedulerImportCheck(SchedulerPlugin):
@@ -62,6 +65,25 @@ def _array_creation(tmpdir):
 
     # Perform a computation using HLG-based array creation
     return da.ones((100,)) + da.zeros((100,))
+
+
+def test_fractional_slice():
+    assert fractional_slice(("x", 4.9), {0: 2}) == (getitem, ("x", 5), (slice(0, 2),))
+
+    assert fractional_slice(("x", 3, 5.1), {0: 2, 1: 3}) == (
+        getitem,
+        ("x", 3, 5),
+        (slice(None, None, None), slice(-3, None)),
+    )
+
+    assert fractional_slice(("x", 2.9, 5.1), {0: 2, 1: 3}) == (
+        getitem,
+        ("x", 3, 5),
+        (slice(0, 2), slice(-3, None)),
+    )
+
+    fs = fractional_slice(("x", 4.9), {0: 2})
+    assert isinstance(fs[1][1], int)
 
 
 def _pq_pyarrow(tmpdir):
